Implement method to restore a loaded game
I haven't really tested it so I'm not sure if this works. The method
restores a game using the following saved classes:
- GameState
- TurnManager
- Map

I believe that these classes contain all the unique game data. There may
be some stuff that I have missed, but the vast majority of game data
should be loaded.